Move resolveWordPressZip to download.ts
