+++ packages/playground/cli/src/cli.ts
@@ -1,7 +1,6 @@
 import fs from 'fs';
 import path from 'path';
 import yargs from 'yargs';
-import { globSync } from 'glob';
 import { startServer } from './server';
 import {
 	PHP,
@@ -23,13 +22,7 @@ import { createNodeFsMountHandler, loadNodeRuntime } from '@php-wasm/node';
 import { RecommendedPHPVersion, zipDirectory } from '@wp-playground/common';
 import { bootWordPress } from '@wp-playground/wordpress';
 import { rootCertificates } from 'tls';
-import {
-	CACHE_FOLDER,
-	fetchSqliteIntegration,
-	fetchWordPress,
-	readAsFile,
-	resolveWPRelease,
-} from './download';
+import { resolveWordPressZip, fetchSqliteIntegration } from './download';
 
 export interface Mount {
 	hostPath: string;
@@ -348,74 +341,6 @@ async function run() {
 			return await requestHandler.request(request);
 		},
 	});
-}
-
-async function resolveWordPressZip(
-	preferredVersion: string,
-	monitor: EmscriptenDownloadMonitor
-): Promise<{ zipFile: File | Promise<File>; localZipPath: string }> {
-	const filenamePrefix = `prebuilt-wp-`;
-	const filenameSuffix = `.zip`;
-	let wpDetails;
-
-	// First, let's grab a pre-installed WordPress zip file we may have
-	// already prepared earlier.
-	const localVersions = globSync(
-		path.join(CACHE_FOLDER, `${filenamePrefix}*${filenameSuffix}`)
-	)
-		.sort()
-		.map((path) => path.split('/').pop()!)
-		.reverse();
-	let resolvedFilename: string | undefined;
-	if (preferredVersion === 'latest') {
-		resolvedFilename = localVersions.filter(
-			(filename) =>
-				!filename.includes('beta') && !filename.includes('nightly')
-		)[0];
-	} else if (preferredVersion === 'beta') {
-		resolvedFilename = localVersions.filter((filename) =>
-			filename.includes('beta')
-		)[0];
-	} else if (preferredVersion === 'nightly') {
-		resolvedFilename = localVersions.filter((filename) =>
-			filename.includes('nightly')
-		)[0];
-	} else {
-		resolvedFilename = localVersions.filter((filename) =>
-			filename.startsWith(`${filenamePrefix}${preferredVersion}`)
-		)[0];
-	}
-	if (resolvedFilename) {
-		const localZipPath = path.join(CACHE_FOLDER, resolvedFilename);
-		const zipFile = readAsFile(localZipPath);
-		return {
-			zipFile,
-			localZipPath,
-		};
-	}
-
-	// We don't have this WordPress version on the device. Let's
-	// do a network lookup.
-	try {
-		wpDetails = await resolveWPRelease(preferredVersion);
-		const localZipPath = path.join(
-			CACHE_FOLDER,
-			`${filenamePrefix}${wpDetails.version}${filenameSuffix}`
-		);
-		return {
-			zipFile: fs.existsSync(localZipPath)
-				? readAsFile(localZipPath)
-				: fetchWordPress(wpDetails.url, monitor),
-			localZipPath,
-		};
-	} catch (e) {
-		throw new Error(
-			`Could not resolve WordPress ${resolvedFilename} from local cache (you're offline)`,
-			{
-				cause: e,
-			}
-		);
-	}
 }
 
 run();
