+++ packages/playground/cli/src/download.ts
@@ -2,6 +2,7 @@ import { EmscriptenDownloadMonitor } from '@php-wasm/progress';
 import { createHash } from 'crypto';
 import fs from 'fs-extra';
 import os from 'os';
+import { globSync } from 'glob';
 import path, { basename } from 'path';
 
 export const CACHE_FOLDER = path.join(os.homedir(), '.wordpress-playground');
@@ -82,6 +83,83 @@ async function downloadTo(
 
 export function readAsFile(path: string, fileName?: string): File {
 	return new File([fs.readFileSync(path)], fileName ?? basename(path));
+}
+
+/**
+ * Offline-first WordPress zip resolution.
+ * It will first look for a pre-installed WordPress zip file in the local cache,
+ * and will only do a network lookup if it doesn't find one.
+ *
+ * @param preferredVersion
+ * @param monitor
+ * @returns
+ */
+export async function resolveWordPressZip(
+	preferredVersion: string,
+	monitor: EmscriptenDownloadMonitor
+): Promise<{ zipFile: File | Promise<File>; localZipPath: string }> {
+	const filenamePrefix = `prebuilt-wp-`;
+	const filenameSuffix = `.zip`;
+	let wpDetails;
+
+	// First, let's grab a pre-installed WordPress zip file we may have
+	// already prepared earlier.
+	const localVersions = globSync(
+		path.join(CACHE_FOLDER, `${filenamePrefix}*${filenameSuffix}`)
+	)
+		.sort()
+		.map((path) => path.split('/').pop()!)
+		.reverse();
+	let resolvedFilename: string | undefined;
+	if (preferredVersion === 'latest') {
+		resolvedFilename = localVersions.filter(
+			(filename) =>
+				!filename.includes('beta') && !filename.includes('nightly')
+		)[0];
+	} else if (preferredVersion === 'beta') {
+		resolvedFilename = localVersions.filter((filename) =>
+			filename.includes('beta')
+		)[0];
+	} else if (preferredVersion === 'nightly') {
+		resolvedFilename = localVersions.filter((filename) =>
+			filename.includes('nightly')
+		)[0];
+	} else {
+		resolvedFilename = localVersions.filter((filename) =>
+			filename.startsWith(`${filenamePrefix}${preferredVersion}`)
+		)[0];
+	}
+	if (resolvedFilename) {
+		const localZipPath = path.join(CACHE_FOLDER, resolvedFilename);
+		const zipFile = readAsFile(localZipPath);
+		return {
+			zipFile,
+			localZipPath,
+		};
+	}
+
+	// We don't have this WordPress version on the device. Let's
+	// do a network lookup.
+	try {
+		wpDetails = await resolveWPRelease(preferredVersion);
+		const localZipPath = path.join(
+			CACHE_FOLDER,
+			`${filenamePrefix}${wpDetails.version}${filenameSuffix}`
+		);
+		return {
+			zipFile: fs.existsSync(localZipPath)
+				? readAsFile(localZipPath)
+				: fetchWordPress(wpDetails.url, monitor),
+			localZipPath,
+		};
+	} catch (e) {
+		throw new Error(
+			`Could not resolve WordPress ${resolvedFilename} from local cache (you're offline)`,
+			{
+				cause: e,
+			}
+		);
+	}
 }
 
 export async function resolveWPRelease(version = 'latest') {
